Add timeout parameter for rcl_wait function
Before this patch, the timeout used for the rcl_wait function is hard
coded which is 10ms. Apparently, this is not reasonable, because it
imposes unnecessary workload on the CPU (We need to lock/unlock the
resources shared between the main thread and the background thread
frequently). Please check the detailed record in the issue.

This patch adds the timeout parameter which developer can customize the
value when starting a node. The default value now is 10ms, negtive value
will block the thread until there is an event.

So you can change your JS code like:
    rclnodejs.spin(node, -1);  // Waits until something happens.

Fix #470